Fix feature payment MoMo and feature Order

+++ Ictshop/Areas/Admin/Controllers/DonhangsController.cs
@@ -13,16 +13,28 @@ namespace Ictshop.Areas.Admin.Controllers
     public class DonhangsController : Controller
     {
         private Qlbanhang db = new Qlbanhang();
+        private SessionController session = new SessionController();
 
         // GET: Admin/Donhangs
         public ActionResult Index()
         {
-            var donhangs = db.Donhangs.Include(d => d.Nguoidung);
+            var u = Session["use"] as Ictshop.Models.Nguoidung;
+            if (session.checkRoleAdmin(u))
+            {
+                return RedirectToRoute("Default", new { controller = "Home", action = "Index" });
+            }
+            var donhangs = db.Donhangs.Include(d => d.Nguoidung).OrderByDescending(n => n.Madon);
             return View(donhangs.ToList());
         }
 
         public ActionResult ChiTietDonHang(int id)
         {
+            var u = Session["use"] as Ictshop.Models.Nguoidung;
+            if (session.checkRoleAdmin(u))
+            {
+                return RedirectToRoute("Default", new { controller = "Home", action = "Index" });
+            }
+
             var donhangs = db.Chitietdonhangs.Where(c => c.Madon == id)
                                             .ToList();
             return View(donhangs);
@@ -31,6 +43,11 @@ namespace Ictshop.Areas.Admin.Controllers
         // GET: Admin/Donhangs/Details/5
         public ActionResult Details(int? id)
         {
+            var u = Session["use"] as Ictshop.Models.Nguoidung;
+            if (session.checkRoleAdmin(u))
+            {
+                return RedirectToRoute("Default", new { controller = "Home", action = "Index" });
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -46,6 +63,11 @@ namespace Ictshop.Areas.Admin.Controllers
         // GET: Admin/Donhangs/Create
         public ActionResult Create()
         {
+            var u = Session["use"] as Ictshop.Models.Nguoidung;
+            if (session.checkRoleAdmin(u))
+            {
+                return RedirectToRoute("Default", new { controller = "Home", action = "Index" });
+            }
             ViewBag.MaNguoidung = new SelectList(db.Nguoidungs, "MaNguoiDung", "Hoten");
             return View();
         }
@@ -57,6 +79,11 @@ namespace Ictshop.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "Madon,Ngaydat,Tinhtrang,MaNguoidung")] Donhang donhang)
         {
+            var u = Session["use"] as Ictshop.Models.Nguoidung;
+            if (session.checkRoleAdmin(u))
+            {
+                return RedirectToRoute("Default", new { controller = "Home", action = "Index" });
+            }
             if (ModelState.IsValid)
             {
                 db.Donhangs.Add(donhang);
@@ -71,6 +98,11 @@ namespace Ictshop.Areas.Admin.Controllers
         // GET: Admin/Donhangs/Edit/5
         public ActionResult Edit(int? id)
         {
+            var u = Session["use"] as Ictshop.Models.Nguoidung;
+            if (session.checkRoleAdmin(u))
+            {
+                return RedirectToRoute("Default", new { controller = "Home", action = "Index" });
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -91,8 +123,14 @@ namespace Ictshop.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "Madon,Ngaydat,Tinhtrang,MaNguoidung")] Donhang donhang)
         {
+            var u = Session["use"] as Ictshop.Models.Nguoidung;
+            if (session.checkRoleAdmin(u))
+            {
+                return RedirectToRoute("Default", new { controller = "Home", action = "Index" });
+            }
             if (ModelState.IsValid)
             {
+                donhang.MaDonMoMo = "DONHANGTRUCTIEP";
                 db.Entry(donhang).State = EntityState.Modified;
                 db.SaveChanges();
                 return RedirectToAction("Index");
@@ -104,6 +142,11 @@ namespace Ictshop.Areas.Admin.Controllers
         // GET: Admin/Donhangs/Delete/5
         public ActionResult Delete(int? id)
         {
+            var u = Session["use"] as Ictshop.Models.Nguoidung;
+            if (session.checkRoleAdmin(u))
+            {
+                return RedirectToRoute("Default", new { controller = "Home", action = "Index" });
+            }
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -121,6 +164,11 @@ namespace Ictshop.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            var u = Session["use"] as Ictshop.Models.Nguoidung;
+            if (session.checkRoleAdmin(u))
+            {
+                return RedirectToRoute("Default", new { controller = "Home", action = "Index" });
+            }
             Donhang donhang = db.Donhangs.Find(id);
             db.Donhangs.Remove(donhang);
             db.SaveChanges();
